Keep workflow tool schemas free of recursive $refs
zod 4's `z.json()` compiles to a self-referential `$defs` entry, and some
providers reject an entire tool list that contains a recursive `$ref`. Every
Grok 4.6 turn failed immediately with a generic provider connection error
because `bb_workflow_run` and `bb_workflow_result` shipped that schema.

Declare the freeform JSON inputs as `unknown` so the wire schema stays flat and
narrow them to `JsonValue` at each call site, which keeps the existing
JSON-compatibility validation. A harness test now fails if any workflow tool
schema regains a `$ref` or `$defs`.
